Fix minor bugs;

+++ app/src/main/java/data/remote/TravelApi.kt
@@ -21,5 +21,5 @@ interface TravelApi {
     suspend fun getBanner(): Response<Unit>
 
     @GET("travels/{travelId}")
-    suspend fun getTravelDetail(@Path("travelId") travelId: Int): Response<ResponseTravelDetailDto>
+    suspend fun getTravelDetail(@Path("travelId") travelId: Long): Response<ResponseTravelDetailDto>
 }
